buggy

+++ pkg/sqlite/database.go
@@ -11,6 +11,7 @@ import (
 	"time"
 
 	"github.com/jmoiron/sqlx"
+	_ "github.com/lib/pq"
 
 	"github.com/stashapp/stash/pkg/fsutil"
 	"github.com/stashapp/stash/pkg/logger"
@@ -36,7 +37,7 @@ const (
 
 var appSchemaVersion uint = 67
 
-//go:embed migrations/*.sql
+//go:embed migrations/*.sql migrationsPostgres/*.sql
 var migrationsBox embed.FS
 
 var (
@@ -81,12 +82,21 @@ type storeRepository struct {
 	Group          *GroupStore
 }
 
+type DatabaseType string
+
+const (
+	PostgresBackend DatabaseType = "POSTGRES"
+	SqliteBackend   DatabaseType = "SQLITE"
+)
+
 type Database struct {
 	*storeRepository
 
-	readDB  *sqlx.DB
-	writeDB *sqlx.DB
-	dbPath  string
+	readDB   *sqlx.DB
+	writeDB  *sqlx.DB
+	dbPath   string
+	dbType   DatabaseType
+	dbString string
 
 	schemaVersion uint
 
@@ -140,16 +150,25 @@ func (db *Database) Ready() error {
 	return nil
 }
 
+func (db *Database) OpenPostgres(dbConnector string) error {
+	db.dbType = PostgresBackend
+	db.dbString = dbConnector
+
+	return db.OpenGeneric()
+}
+
+func (db *Database) OpenSqlite(dbPath string) error {
+	db.dbType = SqliteBackend
+	db.dbPath = dbPath
+
+	return db.OpenGeneric()
+}
+
 // Open initializes the database. If the database is new, then it
 // performs a full migration to the latest schema version. Otherwise, any
 // necessary migrations must be run separately using RunMigrations.
 // Returns true if the database is new.
-func (db *Database) Open(dbPath string) error {
-	db.lock()
-	defer db.unlock()
-
-	db.dbPath = dbPath
-
+func (db *Database) OpenGeneric() error {
 	databaseSchemaVersion, err := db.getDatabaseSchemaVersion()
 	if err != nil {
 		return fmt.Errorf("getting database schema version: %w", err)
@@ -234,26 +253,35 @@ func (db *Database) Close() error {
 	return nil
 }
 
-func (db *Database) open(disableForeignKeys bool, writable bool) (*sqlx.DB, error) {
+func (db *Database) open(disableForeignKeys bool, writable bool) (conn *sqlx.DB, err error) {
+	// Fail-safe
+	err = errors.New("missing backend type")
+
 	// https://github.com/mattn/go-sqlite3
-	url := "file:" + db.dbPath + "?_journal=WAL&_sync=NORMAL&_busy_timeout=50"
-	if !disableForeignKeys {
-		url += "&_fk=true"
+	if db.dbType == SqliteBackend {
+		url := "file:" + db.dbPath + "?_journal=WAL&_sync=NORMAL&_busy_timeout=50"
+		if !disableForeignKeys {
+			url += "&_fk=true"
+		}
+
+		if writable {
+			url += "&_txlock=immediate"
+		} else {
+			url += "&mode=ro"
+		}
+
+		// #5155 - set the cache size if the environment variable is set
+		// default is -2000 which is 2MB
+		if cacheSize := os.Getenv(cacheSizeEnv); cacheSize != "" {
+			url += "&_cache_size=" + cacheSize
+		}
+
+		conn, err = sqlx.Open(sqlite3Driver, url)
+	}
+	if db.dbType == PostgresBackend {
+		conn, err = sqlx.Connect("postgres", db.dbString)
 	}
 
-	if writable {
-		url += "&_txlock=immediate"
-	} else {
-		url += "&mode=ro"
-	}
-
-	// #5155 - set the cache size if the environment variable is set
-	// default is -2000 which is 2MB
-	if cacheSize := os.Getenv(cacheSizeEnv); cacheSize != "" {
-		url += "&_cache_size=" + cacheSize
-	}
-
-	conn, err := sqlx.Open(sqlite3Driver, url)
 	if err != nil {
 		return nil, fmt.Errorf("db.Open(): %w", err)
 	}
@@ -299,6 +327,11 @@ func (db *Database) openWriteDB() error {
 }
 
 func (db *Database) Remove() error {
+	if db.dbType == PostgresBackend {
+		logger.Warn("Postgres backend detected, ignoring Remove request")
+		return nil
+	}
+
 	databasePath := db.dbPath
 	err := db.Close()
 
@@ -326,12 +359,16 @@ func (db *Database) Remove() error {
 }
 
 func (db *Database) Reset() error {
-	databasePath := db.dbPath
+	if db.dbType == PostgresBackend {
+		logger.Warn("Postgres backend detected, ignoring Reset request")
+		return nil
+	}
+
 	if err := db.Remove(); err != nil {
 		return err
 	}
 
-	if err := db.Open(databasePath); err != nil {
+	if err := db.OpenSqlite(db.dbPath); err != nil {
 		return fmt.Errorf("[reset DB] unable to initialize: %w", err)
 	}
 
@@ -341,6 +378,11 @@ func (db *Database) Reset() error {
 // Backup the database. If db is nil, then uses the existing database
 // connection.
 func (db *Database) Backup(backupPath string) (err error) {
+	if db.dbType == PostgresBackend {
+		logger.Warn("Postgres backend detected, ignoring Backup request")
+		return nil
+	}
+
 	thisDB := db.writeDB
 	if thisDB == nil {
 		thisDB, err = sqlx.Connect(sqlite3Driver, "file:"+db.dbPath+"?_fk=true")
@@ -370,6 +412,11 @@ func (db *Database) Anonymise(outPath string) error {
 }
 
 func (db *Database) RestoreFromBackup(backupPath string) error {
+	if db.dbType == PostgresBackend {
+		logger.Warn("Postgres backend detected, ignoring RestoreFromBackup request")
+		return nil
+	}
+
 	logger.Infof("Restoring backup database %s into %s", backupPath, db.dbPath)
 	return os.Rename(backupPath, db.dbPath)
 }
